reference/Books/converter/requirements.txt
# PDF/DOCX to RAG conversion pipeline
pdfplumber>=0.10.0
python-docx>=1.0.0
pyspellchecker>=0.8.0
